requirements.txt
Flask
flask_restful
numpy
tensorflow
requests
pillow
